ItemとUserテーブルを結合してUserGuardsの追加

+++ src/items/items.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { ItemStatus } from './item-status.enum';
 import { Item } from "../entities/item.entity";
 import { CreateItemDto } from './dto/create-item.dto';
 // items.moduleにてitemRepositoryの接続をしているので使える
 import { ItemRepository } from './item.repository';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class ItemsService {
@@ -25,19 +26,27 @@ export class ItemsService {
         return found;
     }
 
-    async create(createItemDto: CreateItemDto): Promise<Item> {
-        return await this.itemRepository.createItem(createItemDto);
+    async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
+        return await this.itemRepository.createItem(createItemDto, user);
     }
 
-    async updateStatus(id: string): Promise<Item> {
+    async updateStatus(id: string, user: User): Promise<Item> {
         const item = await this.findById(id);
+        if (item.userId === user.id) {
+            throw new BadRequestException("自身の商品を購入できません");
+        }
         item.status = ItemStatus.SOLD_OUT;
         item.updatedAt = new Date().toISOString();
         await this.itemRepository.save(item);
         return item;
     }
 
-    async delete(id: string): Promise<void> {
+    async delete(id: string, user: User): Promise<void> {
+        // 削除する投稿
+        const item = await this.findById(id);
+        if (item.userId !== user.id) {
+            throw new BadRequestException("他人の商品は削除できません");
+        }
         await this.itemRepository.delete({ id });
     }
 }
